Rename `accordion` to `accordion-section` to avoid naming collision
The old accordion block will probably need to remain active for as
long as it takes to convert content, so rename `hrswp/accordion`
block to `hrswp/accordion-section` to avoid a naming collision with
the old `hrswp/accordion` block.
