+++ src/blocks/accordions/editor.css
@@ -1,3 +1,18 @@
 .wp-block[data-type="hrswp/accordions"][data-align="wide"] .wp-block {
 	max-width: 100%;
 }
+
+.wp-block-hrswp-accordions {
+
+	& .wp-block-hrswp-accordion-section {
+		border-bottom: 1px solid rgba(0, 0, 0, 0.25);
+	}
+
+	& [data-type="hrswp/accordion-heading"] {
+		background-color: #f7f7f7;
+		border-left: 5px solid var(--accent-crimson);
+		font-size: 1em;
+		line-height: 1.4;
+		padding: 0.5em 0.75em;
+	}
+}
