+++ src/blocks/accordions/variations.js
@@ -25,9 +25,9 @@ const variations = [
 		),
 		isDefault: true,
 		innerBlocks: [
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
 		],
 		scope: [ 'block' ],
 	},
@@ -50,16 +50,16 @@ const variations = [
 			</SVG>
 		),
 		innerBlocks: [
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
-			[ 'hrswp/accordion' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
+			[ 'hrswp/accordion-section' ],
 		],
 		scope: [ 'block' ],
 	},
